requirements.txt
# Function dependencies, for example:
# package>=version
tweepy
gspread
google-cloud-storage
pandas
click==8.0.3
PyYAML==6.0

